feat: added tests directory and refactored accordingly
