+++ tests/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+DATA_DIR = Path(__file__).parent / "data"
